requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
et-xmlfile==1.1.0
idna==3.7
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
